+++ src/pages/Question/components/index.ts
@@ -1,3 +1,3 @@
 export { default as QuestionHeader } from './QuestionHeader/QuestionHeader.tsx';
-export { default as QuestionForm } from './QuestionForm/QuestionForm.tsx';
-export { default as QuestionChat } from './QuestionChat/QuestionChat.tsx';
+export { default as QuestionForm } from './QuestionForm/Form.tsx';
+export { default as QuestionChat } from './QuestionChat/Chat.tsx';
